requirements.txt
gradio
Pillow
torch
numpy
transformers
